Login first during registration before email validation
Without this change, attempting to login on the Registration form with
standard authentication (eg ``ImplicitAuth = Off``) & a validated
account will always result in the "Email Validation Required" message
appearing, preventing login.  This change attempts the login first with
submitted credentials and if that fails (eg they're registering for the
first time), displays the Validation required message.

This was missed in the original PR #500.