Change DC11SchemaArticleAdapter to "proper" format creator's tag
The format of the DC for creator generated for OJS was 'authorname;affiliation', this causes stranges harvests from Vufind or DSpace.

We've had problems with users writing wrongs affiliations and then causing a mess in everyone who harvest from us without considering the different creator tag with ';'

Currently we have a VuFind harvesting a lot of OJS in Brazil (http://oasisbr.ibict.br/vufind/), and they seem to be troubleshooting this by forcing VuFind to get the creator field only before the ';'.

Making OJS export DC with a clean creator tag should make it easier for people to harvest from it.

This is also compliant to DC usage as seem in
http://dublincore.org/documents/usageguide/elements.shtml#creator