pkp/pkp-lib#1536 Fix issue cover image upload
This commit removes the issue cover form and moves the
cover image upload field to the general issue form. It
also removes the issue-specific stylesheet upload
field, and removes a bunch of unused code around the
cover image display logic. It also removes the cover
image caption field, instead relying on the issue
description. Finally, it renames references to the
issue coverPage to coverImage in order to be more
explicit about this being an issue image rather than
a display page. Similar changes will need to be done
for the article cover image in a future commit