Improve error handling for Crossref plugin
At present, https://api.crossref.org/servlet/submissionDownload appears
to be having a technical problem, resulting in 404 responses.

This highlights an issue in the OJS plugin -- Guzzle's HTTP client to
throw an unhandled exception, which means that OJS/PHP return a 500
Internal Server. This in turn doesn't respond with a valid JSON response
when trying to view the status for a "failed" export (e.g. a HTPT
request to a URL like
/$$$call$$$/grid/settings/plugins/settings-plugin-grid/manage?plugin=CrossRefExportPlugin&category=importexport&verb=statusMessage&batchId=123456789&articleId=9999)
-- leading to frontend error "Failed Ajax request or invalid JSON
returned.".

This change configures Guzzle not to throw an exception for status
message retrieval so that the existing non-200 code can operate.
Additionally, the actual server error response body is sent onto the
client, if present, as the original error of "No response from server"
is both unclear and ambiguous as to which server is involved.

This latter error message is reworded to clarify the Crossref server is
the one returning an empty or invalid response.

This same set of changes is also applied to the deposit HTTP calls as
well, to either display the returned error from Crossref or the clearer
invalid response error, if no response body.