pkp/pkp-lib#4235 Don't output invalid chars in OAI setSpec
Valid characters in setSpec [1] are URI unreserved characters [2]
(`a-z`, `A-Z`, `0-9`, `-_.!~*'()`) Currently, OJS creates the setSpec
string by running `urlencode` on journal path and section abbreviation.
This produces invalid character, e.g. `%`, and the OAI-PMH feed fails
to be harvested by clients with strict validation.

This change assumes that journal path is validated to be ASCII
alphanumeric, '-' and '_' elsewhere.

Section abbreviation can be a user supplied containing any Unicode character.
Make it into a valid setSpec string by
- transliterating all non-Latin characters
- stripping all diacritics
- dropping any remaining invalid characters

This will of course not produce a meaningful result for many languages
but it will be a valid setSpec string.

Examples:
"innehåll" -> "innehall"
"Ελληνικά" -> "Ellenika"
"ひらがな" -> "hiragana"

Depends on php-intl extension (which I had to install to get OJS running so should be available?)
Depends on corresponding commit in pkp-lib

[1] https://www.openarchives.org/OAI/2.0/openarchivesprotocol.htm#Set
[2] http://www.ietf.org/rfc/rfc2396.txt